AVANCE

+++ Ejercicios C-Sharp/ProyectoFinal_AIMF/Alumno.cs
@@ -5,7 +5,7 @@ public class Alumno:Persona
     
     public string contra { get; set; }
 
-    public Alumno(int numerocuenta, string nombre)
+    public Alumno(long numerocuenta, string nombre)
     {
         NumeroCuenta = numerocuenta;
         Nombre = nombre;
