+++ Ejercicios C-Sharp/ProyectoFinal_AIMF/Persona.cs
@@ -2,7 +2,7 @@ using System;
 
 public class Persona
 {
-    public int NumeroCuenta { get; set; }
+    public long NumeroCuenta { get; set; }
     public string Nombre { get; set; }
     
 }
